+++ applovin_max/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Versions
 
+## x.x.x
+    * Add support for `AppLovinMAX.setLocationCollectionEnabled(bool enabled)` API
 ## 2.3.1
     * Fix banner/MREC widget ad issue on Android. (https://github.com/AppLovin/AppLovin-MAX-Flutter/issues/32)
 ## 2.3.0